added redirecting to movie details page
