+++ src/movies/moviesPage/MoviePage.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { useParams } from "react-router-dom";
+
+import "./MoviePage.css";
+
+const MoviePage = () => {
+  const { id } = useParams();
+  return <div>MoviesPage</div>;
+};
+
+export default MoviePage;
