+++ src/navbar/NavBar.css
@@ -50,7 +50,7 @@ ul li a {
   text-decoration: none;
   text-align: center;
   color: #ffffffcd;
-  transition: 0.6ms ease-in-out;
+  transition: 0.9ms ease-in-out;
   font-size: 1.2rem;
 }
 
@@ -59,7 +59,7 @@ ul li ul.genres-dropdown li {
   font-size: 1.08rem;
   padding: 1.5px 0;
   cursor: pointer;
-  transition: 0.6ms ease-in-out;
+  transition: 0.9ms ease-in-out;
 }
 
 ul li ul.genres-dropdown li:hover {
