+++ src/resusableComponents/filtering-options/WatchListFilteringOptions.jsx
@@ -1,21 +1,25 @@
 import React from "react";
 
-import "./SortingOptions.css";
+import "../MovieListSortingOptions.css";
 
-const SortingOptions = ({
+const FilteringOptions = ({
   sortingOption,
   setSortingOption,
-  setMovies,
-  setPageNum,
+  setSortedWatchList,
+  watchList,
 }) => {
   const handleOptionChange = (event) => {
     const { name, value } = event.target;
     setSortingOption((prev) => {
       return { ...prev, [name]: value };
     });
+  };
 
-    setMovies([]);
-    setPageNum(1);
+  const handleFilteringOptionsReset = () => {
+    setSortingOption((prev) => {
+      return { ...prev, year: "release_date", rate: "rating" };
+    });
+    setSortedWatchList(watchList);
   };
 
   return (
@@ -32,8 +36,8 @@ const SortingOptions = ({
         value={sortingOption.year}
       >
         <option value="release_date">Year</option>
-        <option value="primary_release_date.asc">Ascending</option>
-        <option value="primary_release_date.desc">Descending</option>
+        <option value="asc">Ascending</option>
+        <option value="desc">Descending</option>
       </select>
       <select
         className="option"
@@ -51,31 +55,14 @@ const SortingOptions = ({
         value={sortingOption.rate}
       >
         <option value="rating">Rating</option>
-        <option value="vote_average.asc">Ascending</option>
-        <option value="vote_average.desc">Descending</option>
+        <option value="asc">Ascending</option>
+        <option value="desc">Descending</option>
       </select>
-      <select
-        className="option"
-        name="services"
-        onChange={(event) => {
-          setSortingOption((prev) => {
-            return {
-              ...prev,
-              rate: "rating",
-              year: "release_date",
-            };
-          });
-          handleOptionChange(event);
-        }}
-        value={sortingOption.services}
-      >
-        <option value="watch_providers">Services</option>
-        <option value="Netflix">Netflix</option>
-        <option value="Apple TV">Apple TV</option>
-        <option value="youtube">YouTube</option>
-      </select>
+      <button className="option" onClick={handleFilteringOptionsReset}>
+        Reset Filters
+      </button>
     </div>
   );
 };
 
-export default SortingOptions;
+export default FilteringOptions;
